+++ packages/svelte/src/config.ts
@@ -1,18 +1,7 @@
-import { CompileOptions } from 'svelte/types/compiler';
 import { PreprocessorGroup } from 'svelte/types/compiler/preprocess';
 
 import { componentTrackingPreprocessor, defaultComponentTrackingOptions } from './preprocessors';
-import { ComponentTrackingInitOptions, SentryPreprocessorGroup } from './types';
-
-export type SvelteConfig = {
-  [key: string]: unknown;
-  preprocess?: PreprocessorGroup[] | PreprocessorGroup;
-  compilerOptions?: CompileOptions;
-};
-
-export type SentrySvelteConfigOptions = {
-  componentTracking?: ComponentTrackingInitOptions;
-};
+import { SentryPreprocessorGroup, SentrySvelteConfigOptions, SvelteConfig } from './types';
 
 const DEFAULT_SENTRY_OPTIONS: SentrySvelteConfigOptions = {
   componentTracking: defaultComponentTrackingOptions,
@@ -36,19 +25,16 @@ export function withSentryConfig(
   };
 
   const originalPreprocessors = getOriginalPreprocessorArray(originalConfig);
-  const sentryPreprocessors: SentryPreprocessorGroup[] = [];
+  const allSentryPreprocessors: SentryPreprocessorGroup[] = [];
 
   const shouldTrackComponents = mergedOptions.componentTracking && mergedOptions.componentTracking.trackComponents;
   if (shouldTrackComponents) {
     // TODO(v8): Remove eslint rule
     // eslint-disable-next-line deprecation/deprecation
-    sentryPreprocessors.push(componentTrackingPreprocessor(mergedOptions.componentTracking));
+    allSentryPreprocessors.push(componentTrackingPreprocessor(mergedOptions.componentTracking));
   }
 
-  const dedupedSentryPreprocessors = sentryPreprocessors.filter(
-    sentryPreproc =>
-      originalPreprocessors.find(p => (p as SentryPreprocessorGroup).id === sentryPreproc.id) === undefined,
-  );
+  const dedupedSentryPreprocessors = dedupePreprocessors(allSentryPreprocessors, originalPreprocessors);
 
   const mergedPreprocessors = [...dedupedSentryPreprocessors, ...originalPreprocessors];
 
@@ -60,7 +46,7 @@ export function withSentryConfig(
 
 /**
  * Standardizes the different ways the user-provided preprocessor option can be specified.
- * Users can specify an array of preprocessors, one single one or nothing at all.
+ * Users can specify an array of preprocessors, a single one or no preprocessor.
  *
  * @param originalConfig the user-provided svelte config oject
  * @return an array of preprocessors or an empty array if no preprocessors were specified
@@ -73,4 +59,14 @@ function getOriginalPreprocessorArray(originalConfig: SvelteConfig): Preprocesso
     return [originalConfig.preprocess];
   }
   return [];
+}
+
+function dedupePreprocessors(
+  sentryPreprocessors: SentryPreprocessorGroup[],
+  originalPreprocessors: PreprocessorGroup[],
+): PreprocessorGroup[] {
+  return sentryPreprocessors.filter(
+    sentryPreproc =>
+      originalPreprocessors.find(p => (p as SentryPreprocessorGroup).id === sentryPreproc.id) === undefined,
+  );
 }
