+++ packages/svelte/src/index.ts
@@ -8,6 +8,7 @@ export * from '@sentry/browser';
 export { init } from './sdk';
 
 // TODO(v8): Remove this export
+// eslint-disable-next-line deprecation/deprecation
 export { componentTrackingPreprocessor } from './preprocessors';
 
 export { trackComponent } from './performance';
